+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0-alpha.23](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.22...v0.8.0-alpha.23) (2021-03-16)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.22](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.21...v0.8.0-alpha.22) (2021-03-15)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.21](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.20...v0.8.0-alpha.21) (2021-03-15)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.20](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.19...v0.8.0-alpha.20) (2021-03-15)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.19](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.18...v0.8.0-alpha.19) (2021-03-15)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.18](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.17...v0.8.0-alpha.18) (2021-03-15)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.17](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.16...v0.8.0-alpha.17) (2021-03-15)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.16](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.15...v0.8.0-alpha.16) (2021-03-12)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.8.0-alpha.15](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.14...v0.8.0-alpha.15) (2021-03-12)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 # [0.8.0-alpha.14](https://bitbucket.org/coveord/ui-kit/compare/v0.8.0-alpha.13...v0.8.0-alpha.14) (2021-03-12)
 
 **Note:** Version bump only for package @coveo/quantic